Move elliptic fluxes initializer
